requirements.txt
python-telegram-bot==20.7
requests==2.31.0
pandas==1.5.3
numpy==1.24.3
flask==2.3.3
gunicorn==21.2.0